Move all classes from controller directory to controllers
