+++ app/src/main/java/edu/kmaooad/capstone23/competences/controllers/UpdateSkillSetController.java
@@ -1,4 +1,4 @@
-package edu.kmaooad.capstone23.competences.controller;
+package edu.kmaooad.capstone23.competences.controllers;
 
 import edu.kmaooad.capstone23.common.TypicalController;
 import edu.kmaooad.capstone23.competences.commands.UpdateSkillSet;
@@ -10,6 +10,6 @@ import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
 
 @Path("/skillset/update")
 @APIResponse(responseCode = "200", content = {
-        @Content(mediaType = "application/json", schema = @Schema(implementation = SkillSetUpdated.class)) })
+        @Content(mediaType = "application/json", schema = @Schema(implementation = SkillSetUpdated.class))})
 public class UpdateSkillSetController extends TypicalController<UpdateSkillSet, SkillSetUpdated> {
 }
